Changing build image to Ubuntu Precise (12.04)
The [builds for Chicago/open311-api-docs](https://travis-ci.org/Chicago/open311-api-docs/builds/263847864) use the same methods as this repo, but do not have the same error. Once difference is that it uses Ubuntu Precise. I am forcing Precise to be used in this build to determine if that fixes the issue.