Equating to Chicago/open311-api-docs
The Chicago/open311-api-docs also uses this same mechanism for deployment to S3. It does not get the same error as this repo. Thus, I've made some changes so the rvm and handling of credentials are the same between the two repos. The only difference _should_ be the `before_deploy` section and, obviously, the `upload-dir` and `on: repo`.